Added struct Queue

+++ CarWash/CarWash/Classes/Employees/Director.swift
@@ -5,7 +5,7 @@ public class Director: Employee {
     // MARK: -
     // MARK: Initializations
     
-    public override init(name: String, gender: Gender, age: Int, salary: Salary = .none, bankAccount: Double = 0.0, money: Int = 0) {
+    public override init(name: String, gender: Gender, age: Int, salary: Salary = .none, bankAccount: Double = 0.0, money: Double = 0) {
         super.init(name: name, gender: gender, age: age, salary: salary, bankAccount: bankAccount, money: money)
     }
