+++ CarWash/CarWash/Protocols/MoneyContanaible.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 public protocol MoneyContainable: AnyObject {
-    var money: Int { get set }
+    var money: Double { get set }
     var bankAccount: Double { get set }
     
     func salaryCoefficient() -> Double
